// src/constants/enums.ts
export enum StudyYear {
  UNKNOWN = 100,
  FOURTH = 4,
  FIFTH = 5,
  SIXTH = 6,
}

export enum StudyBranch {
  MEDICINE = 1,
  DENTISTS = 2,
  PHARMACY = 3,
}

export enum University {
  DAMASCUS = 1,
  TESHREEN = 2,
  BAATH = 3,
  TARTOS = 4,
  ANDALOS = 5,
  FURAT = 6,
  ALEPPO = 7,
  HAMA = 8,
  QALAMON = 9,
  AL_SHAM = 10,
  AL_SORIA = 11,
  NOT_SYR = 100,
}

export enum State {
  ALEPPO = 1,
  DAMASCUS_COUNTRYSIDE = 2,
  DARAA = 3,
  HAMA = 5,
  QUNEITRA = 7,
  LATAKIA = 8,
  TARTUS = 9,
  EASTERN_GOVERNORATES = 10,
  HOMS = 11,
  AS_Suwayda = 12,
  OUT_OF_SYR = 100,
  OTHER = 101,
}

export enum Role {
  USER = "user",
  ADMIN = "admin",
}